+++ build.gradle.kts
@@ -49,7 +49,7 @@ dependencies {
 
     // We can use include() since we only ever use this dependency on fabric
     include(modImplementation("me.lucko:fabric-permissions-api:0.3.1")!!)
-    include(modImplementation("com.github.retrooper:packetevents-fabric:2.5.0")!!)
+//    include(modImplementation("com.github.retrooper:packetevents-fabric:2.5.0")!!)
 
 //    compileOnly("org.spigotmc:spigot:1.18.2-R0.1-SNAPSHOT")
     compileOnly("org.spigotmc:spigot-api:1.18.2-R0.1-SNAPSHOT")
